refactor: size_t

+++ include/lib/string.hpp
@@ -6,18 +6,18 @@
 
 extern "C" {
 void memzero(void* start, void* end);
-void* memcpy(void* dst, const void* src, int n);
-void memset(void* b, int c, int len);
+void* memcpy(void* dst, const void* src, size_t n);
+void memset(void* b, int c, size_t len);
 }
 
-int memcmp(const void* s1, const void* s2, int n);
-int strlen(const char* s);
+int memcmp(const void* s1, const void* s2, size_t n);
+size_t strlen(const char* s);
 char* strcpy(char* dst, const char* src);
 int strcmp(const char* s1, const char* s2);
-int strncmp(const char* s1, const char* s2, int n);
+int strncmp(const char* s1, const char* s2, size_t n);
 const char* strchr(const char* s, char c);
-long strtol(const char* s, const char** endptr = nullptr, int base = 0,
-            int n = 0);
+long strtol(const char* s, const char** endptr = nullptr, size_t base = 0,
+            size_t n = 0);
 
 class string_view {
   const char* buf_;
@@ -108,7 +108,7 @@ class string {
   string& append(const string& o) {
     return append(o.data(), o.size());
   }
-  string& append(const char* s, int n) {
+  string& append(const char* s, size_t n) {
     reserve(size() + n);
     memcpy(end_, s, n);
     end_ += n;
